Library digodoc_common

The Digodoc library scans opam switches and builds maps of them. It is used to build documentation websites, such as https://docs.ocaml.pro

The entry point of this library is the module: Digodoc_common.